se hizo modular el componente de chats, se sigue trabajando en la implementacion de mensajes en tiempo real
