+++ app/src/main/java/io/inzure/app/SplashScreen.kt
@@ -45,7 +45,7 @@ class SplashActivity : ComponentActivity() {
         }
     }
         private fun navigateToLoginView() {
-            startActivity(Intent(this, LoginView::class.java))
+            startActivity(Intent(this, MainActivity::class.java))
             finish()
         }
