+++ local.properties.txt
@@ -1,0 +1,1 @@
+sdk.dir=C:\Users\maxel\AppData\Local\Android\Sdk
